Migrate ClientStats
