+++ packages/services/api/src/modules/operations/resolvers.ts
@@ -4,84 +4,6 @@ import { OperationsModule } from './__generated__/types';
 import { OperationsManager } from './providers/operations-manager';
 
 export const resolvers: OperationsModule.Resolvers = {
-  ClientStats: {
-    totalRequests({ organization, project, target, period, clientName }, _, { injector }) {
-      return injector.get(OperationsManager).countRequestsAndFailures({
-        organization,
-        project,
-        target,
-        period,
-        clients: clientName === 'unknown' ? ['unknown', ''] : [clientName],
-      });
-    },
-    totalVersions({ organization, project, target, period, clientName }, _, { injector }) {
-      return injector.get(OperationsManager).countClientVersions({
-        organization,
-        project,
-        target,
-        period,
-        clientName,
-      });
-    },
-    requestsOverTime(
-      { organization, project, target, period, clientName },
-      { resolution },
-      { injector },
-    ) {
-      return injector.get(OperationsManager).readRequestsOverTime({
-        target,
-        project,
-        organization,
-        period,
-        resolution,
-        clients: clientName === 'unknown' ? ['unknown', ''] : [clientName],
-      });
-    },
-    async operations({ organization, project, target, period, clientName }, args, { injector }) {
-      const operationsManager = injector.get(OperationsManager);
-      const [operations, durations] = await Promise.all([
-        operationsManager.readOperationsStats({
-          organization,
-          project,
-          target,
-          period,
-          clients: clientName === 'unknown' ? ['unknown', ''] : [clientName],
-        }),
-        operationsManager.readDetailedDurationPercentiles({
-          organization,
-          project,
-          target,
-          period,
-          clients: clientName === 'unknown' ? ['unknown', ''] : [clientName],
-        }),
-      ]);
-
-      return operations
-        .map(op => {
-          return {
-            id: hash(`${op.operationName}__${op.operationHash}`),
-            kind: op.kind,
-            name: op.operationName,
-            count: op.count,
-            countOk: op.countOk,
-            percentage: op.percentage,
-            duration: durations.get(op.operationHash)!,
-            operationHash: op.operationHash,
-          };
-        })
-        .sort((a, b) => b.count - a.count);
-    },
-    async versions({ organization, project, target, period, clientName }, { limit }, { injector }) {
-      return injector.get(OperationsManager).readClientVersions({
-        target,
-        project,
-        organization,
-        period,
-        clientName,
-        limit,
-      });
-    },
-  },
   OperationsStats: {
     async operations(
       { organization, project, target, period, operations: operationsFilter, clients },
